character_service/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
django-environ==0.12.0
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.7.0
env==0.1.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
Markdown==3.7
psycopg2-binary==2.9.10
py_eureka_client==0.11.13
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
